Fix DST bug: actually populate day-of-week field
Turns out the clocks were doing daylight savings changes on slightly the
wrong day because they got confused about what day of the week it was.
This is because I'd not been populating the day-of-week field when
setting the RTC.  I'm not sure if I just forgot or were hoping the RTC
peripheral would calculate this for me.

Add a simple function to work out what day of the week it is using
Zeller's algorithm.  Tested by randomly choosing a bunch of dates and
confirming with a calendar that things look right.